split PrometheusMetricsBrowser into small components
